I am enumerating Python modules used in this programming excercise in the README.md
Continuing to fill in bulk_insert.py . Previously we had defined some
constants we needed in that python script.

In this commit I defined the program arguments using the argparse module
and I checked to assure that the input_file was not omitted/invalid,
existed, and was not a directory.